refactor: 將原本的`utils.ts`根據功能性拆分
