+++ src/core/utils/broadcast.ts
@@ -1,0 +1,18 @@
+import { world } from "@minecraft/server";
+
+export namespace BroadcastUtils {
+    export function message(msg: string, type: 'message' | 'actionbar') {
+        for (const player of world.getAllPlayers()) {
+            switch (type) {
+                case 'message': player.sendMessage(msg); break;
+                case 'actionbar': player.onScreenDisplay.setActionBar(msg); break;
+            }
+        }
+    }
+
+    export function sound(soundId: string) {
+        for (const player of world.getAllPlayers()) {
+            player.playSound(soundId);
+        }
+    }
+}
